+++ app/routes/components/Select/changelog.codegen.ts
@@ -1,4 +1,6 @@
 export const changelog = ''
   + '- 2025-09-23: Initial version.\n'
   + '- 2025-09-25: Allow props to accept generics.\n'
+  + "- 2025-09-25: Use `onToggle` instead of `onClick` in `Popover.Trigger`'s render\n"
+  + '  prop.\n'
   ;