+++ src/components/Select/CHANGELOG.md
@@ -1,2 +1,4 @@
 - 2025-09-23: Initial version.
 - 2025-09-25: Allow props to accept generics.
+- 2025-09-25: Use `onToggle` instead of `onClick` in `Popover.Trigger`'s render
+  prop.