+++ src/components/TimeSelector/CHANGELOG.md
@@ -1,2 +1,4 @@
 - 2025-09-11: Initial version.
 - 2025-09-25: Improved TimeSelector styles for better appearance.
+- 2025-09-25: Use `onToggle` instead of `onClick` in `Popover.Trigger`'s render
+  prop.